Use go:embed instead of bin-data to load docs
