+++ internal/command/content.go
@@ -15,9 +15,8 @@
 package command
 
 import (
+	"github.com/mimiro-io/datahub-cli/internal/docs"
 	"os"
-
-	"github.com/mimiro-io/datahub-cli/internal/utils"
 
 	"github.com/mimiro-io/datahub-cli/internal/content"
 	"github.com/pterm/pterm"
@@ -53,7 +52,7 @@ func init() {
 	ContentCmd.SetHelpFunc(func(command *cobra.Command, strings []string) {
 
 		pterm.Println()
-		result := utils.RenderMarkdown(command, "resources/doc-content.md")
+		result := docs.RenderMarkdown(command, "doc-content.md")
 		pterm.Println(result)
 	})
