+++ internal/command/datasets.go
@@ -15,9 +15,8 @@
 package command
 
 import (
+	"github.com/mimiro-io/datahub-cli/internal/docs"
 	"os"
-
-	"github.com/mimiro-io/datahub-cli/internal/utils"
 
 	"github.com/mimiro-io/datahub-cli/internal/datasets"
 	"github.com/pterm/pterm"
@@ -57,7 +56,7 @@ func init() {
 
 	DatasetCmd.SetHelpFunc(func(command *cobra.Command, strings []string) {
 		pterm.Println()
-		result := utils.RenderMarkdown(command, "resources/doc-dataset.md")
+		result := docs.RenderMarkdown(command, "doc-dataset.md")
 		pterm.Println(result)
 	})
 }
