+++ internal/command/transform.go
@@ -15,9 +15,8 @@
 package command
 
 import (
+	"github.com/mimiro-io/datahub-cli/internal/docs"
 	"os"
-
-	"github.com/mimiro-io/datahub-cli/internal/utils"
 
 	"github.com/mimiro-io/datahub-cli/internal/transform"
 	"github.com/pterm/pterm"
@@ -48,7 +47,7 @@ func init() {
 
 	TransformCmd.SetHelpFunc(func(command *cobra.Command, strings []string) {
 		pterm.Println()
-		result := utils.RenderMarkdown(command, "resources/doc-transform.md")
+		result := docs.RenderMarkdown(command, "doc-transform.md")
 		pterm.Println(result)
 	})
 }
